FIX - app could crash if default translation not exists
We tried to find a unknownError by default but this one is an extra translation I've personnaly add. Plus you don't necessary try to send an error to the client. The best thing to do is to send the message you try to translate directly to the client if no translation exists.